requirements.txt
aiohttp==3.7.3
aioreloader==0.4.0
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
idna==2.10
motor==2.3.0
multidict==5.1.0
pymongo==3.11.2
PyYAML==5.3.1
trafaret==2.1.0
typing-extensions==3.7.4.3
uvloop==0.14.0
yarl==1.6.3
